\name{royston-package}
\alias{royston-package}
\alias{royston}
\docType{package}
\title{
Royston's Multivariate Normality Test
}
\description{
Performs a multivariate normality test based on Royston's H test
}
\details{
\tabular{ll}{
Package: \tab royston\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
royston.test(a)
}
\author{
Selcuk Korkmaz
Maintainer: Selcuk Korkmaz <selcuk.korkmaz@hacettepe.edu.tr>
}
